Remove HttpRequestFactory.
